requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.18.0
pandas>=2.0.0
openpyxl>=3.1.0
